failing test case
git-svn-id: https://svn.apache.org/repos/asf/incubator/directory/eve/trunk@55750 13f79535-47bb-0310-9956-ffa450edef68

+++ backend/core/src/test/org/apache/eve/jndi/AbstractJndiTest.java
@@ -38,6 +38,9 @@ public class AbstractJndiTest extends TestCase
     /** the context root for the system partition */
     protected LdapContext sysRoot;
 
+    /** flag whether to delete database files for each test or not */
+    protected boolean doDelete = true;
+
 
     /**
      * Get's the initial context factory for the provider's ou=system context
@@ -81,7 +84,11 @@ public class AbstractJndiTest extends TestCase
         }
 
         sysRoot = null;
-        File file = new File( "target/eve" );
-        FileUtils.deleteDirectory( file );
+
+        if ( doDelete == true )
+        {
+            File file = new File( "target/eve" );
+            FileUtils.deleteDirectory( file );
+        }
     }
 }
